Move memcpy operation to its own file.
