+++ jlm/llvm/ir/operators.hpp
@@ -14,6 +14,7 @@
 #include <jlm/llvm/ir/operators/lambda.hpp>
 #include <jlm/llvm/ir/operators/load.hpp>
 #include <jlm/llvm/ir/operators/LoadVolatile.hpp>
+#include <jlm/llvm/ir/operators/MemCpy.hpp>
 #include <jlm/llvm/ir/operators/operators.hpp>
 #include <jlm/llvm/ir/operators/Phi.hpp>
 #include <jlm/llvm/ir/operators/sext.hpp>
