+++ jlm/llvm/ir/operators/operators.cpp
@@ -1273,30 +1273,4 @@ FreeOperation::copy() const
   return std::unique_ptr<rvsdg::operation>(new FreeOperation(*this));
 }
 
-/* memcpy operator */
-
-Memcpy::~Memcpy()
-{}
-
-bool
-Memcpy::operator==(const operation & other) const noexcept
-{
-  /*
-    Avoid CNE for memcpy operator
-  */
-  return this == &other;
-}
-
-std::string
-Memcpy::debug_string() const
-{
-  return "Memcpy";
-}
-
-std::unique_ptr<rvsdg::operation>
-Memcpy::copy() const
-{
-  return std::unique_ptr<rvsdg::operation>(new Memcpy(*this));
-}
-
 }
