+++ jlm/llvm/ir/operators/operators.hpp
@@ -2594,88 +2594,6 @@ private:
   }
 };
 
-/* memcpy operation */
-
-class Memcpy final : public jlm::rvsdg::simple_op
-{
-public:
-  virtual ~Memcpy();
-
-  Memcpy(
-      const std::vector<jlm::rvsdg::port> & operandPorts,
-      const std::vector<jlm::rvsdg::port> & resultPorts)
-      : simple_op(operandPorts, resultPorts)
-  {}
-
-  virtual bool
-  operator==(const operation & other) const noexcept override;
-
-  virtual std::string
-  debug_string() const override;
-
-  virtual std::unique_ptr<jlm::rvsdg::operation>
-  copy() const override;
-
-  static std::unique_ptr<llvm::tac>
-  create(
-      const variable * destination,
-      const variable * source,
-      const variable * length,
-      const variable * isVolatile,
-      const std::vector<const variable *> & memoryStates)
-  {
-    auto operandPorts = CheckAndCreateOperandPorts(length->type(), memoryStates.size());
-    auto resultPorts = CreateResultPorts(memoryStates.size());
-
-    std::vector<const variable *> operands = { destination, source, length, isVolatile };
-    operands.insert(operands.end(), memoryStates.begin(), memoryStates.end());
-
-    Memcpy op(operandPorts, resultPorts);
-    return tac::create(op, operands);
-  }
-
-  static std::vector<jlm::rvsdg::output *>
-  create(
-      jlm::rvsdg::output * destination,
-      jlm::rvsdg::output * source,
-      jlm::rvsdg::output * length,
-      jlm::rvsdg::output * isVolatile,
-      const std::vector<jlm::rvsdg::output *> & memoryStates)
-  {
-    auto operandPorts = CheckAndCreateOperandPorts(length->type(), memoryStates.size());
-    auto resultPorts = CreateResultPorts(memoryStates.size());
-
-    std::vector<jlm::rvsdg::output *> operands = { destination, source, length, isVolatile };
-    operands.insert(operands.end(), memoryStates.begin(), memoryStates.end());
-
-    Memcpy op(operandPorts, resultPorts);
-    return jlm::rvsdg::simple_node::create_normalized(destination->region(), op, operands);
-  }
-
-private:
-  static std::vector<jlm::rvsdg::port>
-  CheckAndCreateOperandPorts(const jlm::rvsdg::type & length, size_t nMemoryStates)
-  {
-    if (length != jlm::rvsdg::bit32 && length != jlm::rvsdg::bit64)
-      throw jlm::util::error("Expected 32 bit or 64 bit integer type.");
-
-    if (nMemoryStates == 0)
-      throw jlm::util::error("Number of memory states cannot be zero.");
-
-    PointerType pointerType;
-    std::vector<jlm::rvsdg::port> ports = { pointerType, pointerType, length, jlm::rvsdg::bit1 };
-    ports.insert(ports.end(), nMemoryStates, { MemoryStateType::Create() });
-
-    return ports;
-  }
-
-  static std::vector<jlm::rvsdg::port>
-  CreateResultPorts(size_t nMemoryStates)
-  {
-    return std::vector<jlm::rvsdg::port>(nMemoryStates, { MemoryStateType::Create() });
-  }
-};
-
 /*
   FIXME: This function should be in librvsdg and not in libllvm.
 */
